+++ src/pages/cart/CartContent/OrderContent/CartList/CartItem/ProductImage/ProductImage.styled.ts
@@ -1,0 +1,11 @@
+import styled from "@emotion/styled";
+
+export const Image = styled.img`
+  width: 112px;
+  height: 112px;
+  border-radius: 8px;
+  object-fit: cover;
+  object-position: center;
+  display: block;
+  flex-shrink: 0;
+`;
